fix(eggs): eggs padding; plus moved hook to lib
